requirements.txt
anyio==4.3.0
certifi==2024.2.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
pip-autoremove==0.10.0
python-dotenv==1.0.1
python-telegram-bot==21.0.1
setuptools==69.2.0
sniffio==1.3.1
SQLAlchemy==2.0.29
typing_extensions==4.10.0
